refactor(core): split canvas mark drawers and scale-training tests
Extract focus-mask helpers and point/path/segment drawers from the
canvas-marks module; keep drawBatch dispatch as the thin orchestrator
with stable public re-exports.

Replace the 434-line pipeline-scale-training monolith with focused suites
under tests/pipeline-scale-training/ (isBarLike, trainAxis, color scale,
axis inputs, runPipeline anchors, force-zero warnings) sharing fixtures.
