+++ packages/core/src/dom/canvas-marks-points.ts
@@ -1,0 +1,128 @@
+/// <reference lib="dom" />
+/**
+ * Canvas point (and point-subset) drawers.
+ */
+import type { PointsBatch } from "../scene.js";
+import type { ThemeTokens } from "../theme.js";
+import { themeVar } from "../theme.js";
+import type { ColorResolver } from "./canvas-dom.js";
+import { maskIncludes, type PrimitiveFocusMask } from "./canvas-marks-mask.js";
+
+function tracePoint(ctx: CanvasRenderingContext2D, batch: PointsBatch, j: number): void {
+  const x = batch.positions[j * 2]!;
+  const y = batch.positions[j * 2 + 1]!;
+  const size = batch.size;
+  switch (batch.shape) {
+    case "square":
+      ctx.rect(x - size, y - size, size * 2, size * 2);
+      break;
+    case "triangle":
+      // Same proportions as the SVG renderer's triangle path.
+      ctx.moveTo(x, y - size * 1.2);
+      ctx.lineTo(x + size * 1.1, y + size * 0.9);
+      ctx.lineTo(x - size * 1.1, y + size * 0.9);
+      ctx.closePath();
+      break;
+    default:
+      ctx.moveTo(x + size, y);
+      ctx.arc(x, y, size, 0, Math.PI * 2);
+  }
+}
+
+export function drawPoints(
+  ctx: CanvasRenderingContext2D,
+  batch: PointsBatch,
+  theme: ThemeTokens,
+  resolve: ColorResolver,
+): void {
+  const themeInk = resolve(themeVar("ink", theme));
+  const n = batch.rowIndex.length;
+  if (batch.colors === undefined) {
+    // Single fill: one path for the whole batch (the fast path that makes
+    // canvas worth it at high counts).
+    ctx.fillStyle = batch.fill === null ? themeInk : resolve(batch.fill);
+    ctx.beginPath();
+    for (let j = 0; j < n; j++) tracePoint(ctx, batch, j);
+    ctx.fill();
+    return;
+  }
+  // Per-mark colors: batch consecutive same-color runs.
+  let runStart = 0;
+  while (runStart < n) {
+    const color = batch.colors[runStart] ?? batch.fill ?? themeInk;
+    let runEnd = runStart + 1;
+    while (runEnd < n && (batch.colors[runEnd] ?? batch.fill ?? themeInk) === color) runEnd++;
+    ctx.fillStyle = resolve(color);
+    ctx.beginPath();
+    for (let j = runStart; j < runEnd; j++) tracePoint(ctx, batch, j);
+    ctx.fill();
+    runStart = runEnd;
+  }
+}
+
+export function drawPointsSubset(
+  ctx: CanvasRenderingContext2D,
+  batch: PointsBatch,
+  theme: ThemeTokens,
+  resolve: ColorResolver,
+  mask: PrimitiveFocusMask,
+  focused: boolean,
+): void {
+  const includes = (index: number) => maskIncludes(mask, index) === focused;
+  const themeInk = resolve(themeVar("ink", theme));
+  const n = batch.rowIndex.length;
+  if (batch.colors === undefined) {
+    ctx.fillStyle = batch.fill === null ? themeInk : resolve(batch.fill);
+    ctx.beginPath();
+    let traced = false;
+    for (let j = 0; j < n; j++) {
+      if (!includes(j)) continue;
+      tracePoint(ctx, batch, j);
+      traced = true;
+    }
+    if (traced) ctx.fill();
+    return;
+  }
+  // Interactive masks must not turn alternating categorical colors into one
+  // beginPath/fill pair per point. For the normal small categorical case,
+  // batch by first-seen color while retaining the focused-over-muted pass
+  // boundary. Fall back to contiguous runs for high-cardinality colors.
+  const uniqueColors: string[] = [];
+  const seenColors = new Set<string>();
+  for (let j = 0; j < n && uniqueColors.length <= 64; j++) {
+    const color = batch.colors[j] ?? batch.fill ?? themeInk;
+    if (seenColors.has(color)) continue;
+    seenColors.add(color);
+    uniqueColors.push(color);
+  }
+  if (uniqueColors.length <= 64) {
+    for (const color of uniqueColors) {
+      ctx.fillStyle = resolve(color);
+      ctx.beginPath();
+      let traced = false;
+      for (let j = 0; j < n; j++) {
+        if ((batch.colors[j] ?? batch.fill ?? themeInk) !== color || !includes(j)) continue;
+        tracePoint(ctx, batch, j);
+        traced = true;
+      }
+      if (traced) ctx.fill();
+    }
+    return;
+  }
+  let runStart = 0;
+  while (runStart < n) {
+    const color = batch.colors[runStart] ?? batch.fill ?? themeInk;
+    let runEnd = runStart + 1;
+    while (runEnd < n && (batch.colors[runEnd] ?? batch.fill ?? themeInk) === color) runEnd++;
+    ctx.fillStyle = resolve(color);
+    ctx.beginPath();
+    let traced = false;
+    for (let j = runStart; j < runEnd; j++) {
+      if (!includes(j)) continue;
+      tracePoint(ctx, batch, j);
+      traced = true;
+    }
+    if (traced) ctx.fill();
+    runStart = runEnd;
+  }
+}
